third_party/WebKit/LayoutTests/media/video-enter-fullscreen-without-user-gesture-expected.txt

Test that video webkitRequestFullscreen() works without any user gesture if the requirement is removed.

EVENT(canplaythrough)

* No user gesture initiated
EXPECTED (video.paused == 'true') OK
RUN(video.webkitRequestFullscreen())
EVENT(webkitfullscreenchange)
END OF TEST
